pyproject.toml
[tool.poetry]
name = "Burrito"
version = "2.0.0.alpha"
description = "API for the issue tracker"
authors = ["DimonBor", "m-o-d-e-r"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
anyio = "^3.6.2"
distro = "^1.8.0"
fastapi = "^0.111.0"
h11 = "^0.14.0"
idna = "^3.4"
peewee = "^3.16.0"
pydantic = "^1.10.7"
sniffio = "^1.3.0"
starlette = "^0.37.2"
typing-extensions = "^4.5.0"
uvicorn = "^0.29.0"
requests = "^2.28.2"
pymysql = {extras = ["rsa"], version = "^1.0.3"}
python-dotenv = "^1.0.0"
python-multipart = "^0.0.9"
aiofiles = "^23.1.0"
cryptography = "^42.0.5"
tabulate = "^0.9.0"
redis = "^5.0.1"
pymongo = "^4.4.1"
schedule = "^1.2.0"
pytz = "^2024.1"
argon2-cffi = "^23.1.0"
orjson = "^3.9.7"
websockets = "^12.0"
funcy = "^2.0"
deptry = "^0.16.1"
pyjwt = "^2.8.0"
httpx = "^0.27.0"
apache-airflow = {version = "2.9.1", extras = ["celery"]}
psycopg2-binary = "^2.9.9"


[tool.poetry.group.dev.dependencies]
prospector = "^1.10.2"

[tool.isort]
profile = "black"
known_local_folder = "burrito"
sections = "FUTURE,STDLIB,THIRDPARTY,LOCALFOLDER"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"